+++ CHANGELOG.md
@@ -664,6 +664,13 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 - Split with `.`
 - `1.0` = `1`
 
+## 0166 - Fraction to Recurring Decimal
+
+- Integer
+- Finite decimal
+- Recurring decimal
+- Nagtive
+
 ## 0167 - Two Sum II - Input array is sorted
 
 - Map